+++ src/main/java/com/example/umc8th/domain/reply/service/query/ReplyQueryService.java
@@ -1,0 +1,10 @@
+package com.example.umc8th.domain.reply.service.query;
+
+import com.example.umc8th.domain.reply.entity.Reply;
+
+import java.util.List;
+
+public interface ReplyQueryService {
+    Reply getReply(Long id);
+    List<Reply> getRepliesByArticleId(Long articleId);
+}
