--- src/main/java/com/example/umc8th/reply/service/command/ReplyCommandServiceImpl.java
@@ -1,34 +1,0 @@
-package com.example.umc8th.reply.service.command;
-
-import com.example.umc8th.article.entity.Article;
-import com.example.umc8th.article.repository.ArticleRepository;
-import com.example.umc8th.global.apiPayload.code.ArticleErrorCode;
-import com.example.umc8th.global.apiPayload.exception.GeneralException;
-import com.example.umc8th.reply.dto.ReplyRequestDTO;
-import com.example.umc8th.reply.entity.Reply;
-import com.example.umc8th.reply.repository.ReplyRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@Transactional
-@RequiredArgsConstructor
-public class ReplyCommandServiceImpl implements ReplyCommandService {
-
-    private final ReplyRepository replyRepository;
-    private final ArticleRepository articleRepository;
-
-    @Override
-    public Reply createReply(ReplyRequestDTO.CreateReplyDTO dto, Long articleId) {
-        Article article = articleRepository.findById(articleId)
-                .orElseThrow(() -> new GeneralException(ArticleErrorCode.ARTICLE_NOT_FOUND));
-
-        return replyRepository.save(
-                Reply.builder()
-                        .content(dto.getContent())
-                        .article(article)
-                        .build()
-        );
-    }
-}
